requirements.txt
Flask==3.0.0
Markdown==3.5.2

google-generativeai==0.3.2
python-dotenv==1.0.0

langchain-google-genai==0.0.6
langchain==0.1.0
pypdf==3.17.4
faiss-cpu==1.7.4